Added observed generation to status
As desired state may flip between primary and secondary
added an ObservedGeneration to status to reflect which
generation of the resource is the status updated for.

This can assist is ensuring that the desired state in
the latest generation of the resource is what is
reconciled and its status can be used for further
operations.

Signed-off-by: Shyamsundar Ranganathan <srangana@redhat.com>